Refactor the api mock creator out.
